# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PexCoin Trading Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
  /** @nullable */
  phone?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserBalance {
  usdt: number;
  btc: number;
  eth: number;
}

export interface CryptoPrice {
  symbol: string;
  name: string;
  price: number;
  change: number;
  /** @nullable */
  iconUrl?: string | null;
}

export interface TickerItem {
  pair: string;
  price: number;
  change: number;
}

export type CreateTransactionBodyType =
  (typeof CreateTransactionBodyType)[keyof typeof CreateTransactionBodyType];

export const CreateTransactionBodyType = {
  deposit: "deposit",
  withdrawal: "withdrawal",
} as const;

export interface CreateTransactionBody {
  type: CreateTransactionBodyType;
  amount: number;
  currency: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  note?: string | null;
}

export interface Transaction {
  id: number;
  type: string;
  amount: number;
  currency: string;
  status: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface AdminAuthResponse {
  token: string;
  role: string;
}

export interface AdminUser {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  status: string;
  usdtBalance: number;
  btcBalance: number;
  ethBalance: number;
  createdAt: string;
}

export interface AdminUpdateUserBody {
  /** @nullable */
  status?: string | null;
  /** @nullable */
  usdtBalance?: number | null;
  /** @nullable */
  btcBalance?: number | null;
  /** @nullable */
  ethBalance?: number | null;
  /** @nullable */
  role?: string | null;
}

export interface AdminTransaction {
  id: number;
  userId: number;
  userEmail: string;
  type: string;
  amount: number;
  currency: string;
  status: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export interface AdminStats {
  totalUsers: number;
  totalDeposits: number;
  totalWithdrawals: number;
  pendingTransactions: number;
  activeUsers: number;
}