# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PexCoin Trading Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: users
    description: User management
  - name: crypto
    description: Crypto market data
  - name: transactions
    description: Deposit and withdrawal transactions
  - name: admin
    description: Admin panel operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /crypto/prices:
    get:
      operationId: getCryptoPrices
      tags: [crypto]
      summary: Get all crypto pair prices
      responses:
        "200":
          description: List of crypto prices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CryptoPrice"

  /crypto/ticker:
    get:
      operationId: getMarketTicker
      tags: [crypto]
      summary: Get top market ticker data (BTC, ETH, BCH)
      responses:
        "200":
          description: Ticker data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TickerItem"

  /transactions:
    get:
      operationId: getMyTransactions
      tags: [transactions]
      summary: Get current user transactions
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a deposit or withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/balance:
    get:
      operationId: getMyBalance
      tags: [users]
      summary: Get current user balance
      responses:
        "200":
          description: User balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserBalance"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Admin logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: Admin get all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"

  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Admin get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Admin update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"

  /admin/transactions:
    get:
      operationId: adminGetTransactions
      tags: [admin]
      summary: Admin get all transactions
      responses:
        "200":
          description: All transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminTransaction"

  /admin/transactions/{id}/approve:
    post:
      operationId: adminApproveTransaction
      tags: [admin]
      summary: Admin approve a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTransaction"

  /admin/transactions/{id}/reject:
    post:
      operationId: adminRejectTransaction
      tags: [admin]
      summary: Admin reject a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTransaction"

  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Admin dashboard stats
      responses:
        "200":
          description: Platform statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
      required:
        - email
        - password
        - name

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - createdAt

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token

    UserBalance:
      type: object
      properties:
        usdt:
          type: number
        btc:
          type: number
        eth:
          type: number
      required:
        - usdt
        - btc
        - eth

    CryptoPrice:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        price:
          type: number
        change:
          type: number
        iconUrl:
          type: ["string", "null"]
      required:
        - symbol
        - name
        - price
        - change

    TickerItem:
      type: object
      properties:
        pair:
          type: string
        price:
          type: number
        change:
          type: number
      required:
        - pair
        - price
        - change

    CreateTransactionBody:
      type: object
      properties:
        type:
          type: string
          enum: [deposit, withdrawal]
        amount:
          type: number
        currency:
          type: string
        address:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
      required:
        - type
        - amount
        - currency

    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
        address:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - type
        - amount
        - currency
        - status
        - createdAt

    AdminLoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    AdminAuthResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
      required:
        - token
        - role

    AdminUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        status:
          type: string
        usdtBalance:
          type: number
        btcBalance:
          type: number
        ethBalance:
          type: number
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - status
        - usdtBalance
        - btcBalance
        - ethBalance
        - createdAt

    AdminUpdateUserBody:
      type: object
      properties:
        status:
          type: ["string", "null"]
        usdtBalance:
          type: ["number", "null"]
        btcBalance:
          type: ["number", "null"]
        ethBalance:
          type: ["number", "null"]
        role:
          type: ["string", "null"]

    AdminTransaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userEmail:
          type: string
        type:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
        address:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - userEmail
        - type
        - amount
        - currency
        - status
        - createdAt

    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalDeposits:
          type: number
        totalWithdrawals:
          type: number
        pendingTransactions:
          type: integer
        activeUsers:
          type: integer
      required:
        - totalUsers
        - totalDeposits
        - totalWithdrawals
        - pendingTransactions
        - activeUsers