# lib/api-zod/src/generated/types/adminTransaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PexCoin Trading Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminTransaction {
  id: number;
  userId: number;
  userEmail: string;
  type: string;
  amount: number;
  currency: string;
  status: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}