# lib/api-zod/src/generated/types/adminUpdateUserBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PexCoin Trading Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUpdateUserBody {
  /** @nullable */
  status?: string | null;
  /** @nullable */
  usdtBalance?: number | null;
  /** @nullable */
  btcBalance?: number | null;
  /** @nullable */
  ethBalance?: number | null;
  /** @nullable */
  role?: string | null;
}