# lib/api-zod/src/generated/types/createTransactionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PexCoin Trading Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTransactionBodyType } from "./createTransactionBodyType";

export interface CreateTransactionBody {
  type: CreateTransactionBodyType;
  amount: number;
  currency: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  note?: string | null;
}