# lib/api-zod/src/generated/types/tickerItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PexCoin Trading Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TickerItem {
  pair: string;
  price: number;
  change: number;
}