# lib/api-zod/src/generated/types/userBalance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PexCoin Trading Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserBalance {
  usdt: number;
  btc: number;
  eth: number;
}